+++ examples/calc_pedestals.py
@@ -41,12 +41,12 @@ if __name__ == "__main__":
     # loop over all events, all telescopes and all channels and call
     # the calc_peds function defined above to do some work:
     for event in event_source(filename):
-        for telid in event.r0.tels_with_data:
-            for chan in range(event.r0.tel[telid].waveform.shape[0]):
+        for tel_id in event.r0.tels_with_data:
+            for chan in range(event.r0.tel[tel_id].waveform.shape[0]):
 
-                print(f"CT{telid} chan {chan}:")
+                print(f"CT{tel_id} chan {chan}:")
 
-                traces = event.r0.tel[telid].waveform[chan, ...]
+                traces = event.r0.tel[tel_id].waveform[chan, ...]
 
                 # skip telescopes without timeseries data
                 if traces.shape[1] == 1: